Create: Validator - LivroRepository - Exception - TryCatch

+++ src/main/java/io/github/alanpcavalcante/libraryapi/service/AutorService.java
@@ -1,10 +1,12 @@
 package io.github.alanpcavalcante.libraryapi.service;
 
 
-import io.github.alanpcavalcante.libraryapi.exception.AutorDuplicado;
-import io.github.alanpcavalcante.libraryapi.exception.NenhumAutorEncontradoException;
+
+import io.github.alanpcavalcante.libraryapi.exceptions.OperacaoNaoPermitidaException;
 import io.github.alanpcavalcante.libraryapi.model.Autor;
 import io.github.alanpcavalcante.libraryapi.repository.AutorRepository;
+import io.github.alanpcavalcante.libraryapi.repository.LivroRepository;
+import io.github.alanpcavalcante.libraryapi.validator.AutorValidator;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
@@ -16,12 +18,21 @@ import java.util.UUID;
 public class AutorService {
 
     public final AutorRepository repository;
+    public final AutorValidator validator;
+    public final LivroRepository livroRepository;
 
-    public AutorService(AutorRepository repository) {
+    public AutorService(
+            AutorRepository repository,
+            AutorValidator validator,
+            LivroRepository livroRepository) {
         this.repository = repository;
+        this.validator = validator;
+        this.livroRepository = livroRepository;
     }
 
-    public List<Autor> autoresRepetidos(Autor autor) {
+
+
+    public Optional<Autor> autoresRepetidos(Autor autor) {
         return repository.findByNomeAndDataNascimentoAndNacionalidade(
                 autor.getNome(),
                 autor.getDataNascimento(),
@@ -29,26 +40,40 @@ public class AutorService {
         );
     }
 
+
+
     public List<Autor> todosAutores() {
         return repository.findAll();
     }
+
+
 
     public Optional<Autor> pegarAutorPorId(String id) {
         return repository.findById(UUID.fromString(id));
     }
 
+
+
     public Autor salvar(Autor autor) {
+        validator.validar(autor);
         return repository.save(autor);
     }
 
-    public Autor atualizarAutor(String id, Autor autorAtualizado) {
-
-    }
 
 
     public void deletarAutor(Autor autor) {
+        if (possuiLivro(autor)) {
+            throw new OperacaoNaoPermitidaException(
+                    "Não é permitido excluir um autor que possui livros cadastrados!"
+            );
+        }
         repository.delete(autor);
     }
 
+
+
+    public boolean possuiLivro(Autor autor) {
+        return livroRepository.existsByAutor(autor);
+    }
 
 }
